# Conflicts:
#   android/Signal/app/src/main/java/com/ongo/signal/di/NetworkModule.kt

+++ android/Signal/app/src/main/java/com/ongo/signal/data/repository/chat/ChatUseCases.kt
@@ -6,11 +6,11 @@ import com.ongo.signal.data.model.chat.ChatHomeDTO
 interface ChatUseCases {
     suspend fun loadChats(): List<ChatHomeDTO>
     suspend fun saveChat(room: ChatHomeDTO)
-    suspend fun loadDetailList(id: Int): List<ChatHomeChildDto>
-    suspend fun saveDetailList(message: ChatHomeChildDto, id: Int)
+    suspend fun loadDetailList(id: Long): List<ChatHomeChildDto>
+    suspend fun saveDetailList(message: ChatHomeChildDto, id: Long)
     fun timeSetting(): String
     suspend fun stompSend(item: ChatHomeChildDto)
-    suspend fun stompGet(chatRoomNumber: Int, onSuccess: (Int) -> Unit)
-    suspend fun connectedWebSocket(chatRoomNumber: Int)
+    suspend fun stompGet(chatRoomNumber: Long, onSuccess: (Long) -> Unit)
+    suspend fun connectedWebSocket(chatRoomNumber: Long)
     suspend fun stompDisconnect()
 }
